requirement.txt
python-socketio
eventlet
websocket-client